requirements.txt
# Deep Learning
tensorflow==2.15.0

# Image Processing
opencv-python==4.8.0.76
Pillow==9.5.0
pydicom==2.4.3

# Data & ML Utilities
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.7.3
plotly==5.17.0

# Backend
flask==3.0.0
flask-cors==4.0.0

# Groq API
groq==0.4.2

# Explainable AI
tf-keras-vis==0.8.5

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0